requirements.txt
playwright==1.46.0
python-dotenv==1.0.1
APScheduler==3.10.4
pydantic==2.9.2
beautifulsoup4==4.12.3
lxml==5.3.0
